requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0